+++ CHANGES/plugin_api/6514.removal
@@ -1,0 +1,1 @@
+Rehomed QueryModelResource to pulpcore.plugin.importexport.
